+++ src/components/App.js
@@ -58,6 +58,7 @@ const appReducer = (state, action) => {
       };
 
     case ACTIONS.EDIT_INGREDIENT:
+      console.log("editing ingredient");
       const { content, id, colId } = action.payload;
       const tempArr = state.dndColumns[colId].itemsArr; // make a copy of column's item array
       var result = tempArr.filter((obj) => obj.id === id); // sift out id that matches obj id within array
@@ -75,6 +76,7 @@ const appReducer = (state, action) => {
         ...state,
         activePage: action.payload,
       };
+
     default:
       return state;
   }
