+++ src/pages/Recipes.js
@@ -1,101 +1,53 @@
-import React from "react";
-import complexData from "../data/complexData";
-import {
-  Button,
-  Card,
-  CardContent,
-  CardMedia,
-  Container,
-  Grid,
-  Typography,
-} from "@material-ui/core";
-import { makeStyles } from "@material-ui/core/styles";
+import React, { useEffect, useState, useContext } from "react";
+import { DataContext } from "../components/App";
+import RecipeCards from "../components/Cards/RecipeCards";
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-    marginTop: "5vh",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  grid: {
-    flexGrow: 1,
-  },
-  btn: {
-    marginTop: "1vh",
-  },
-}));
+const Recipes = (props) => {
+  const value = useContext(DataContext);
+  const { state } = value;
 
-const handleClick = () => {};
+  const [data, setData] = useState(null);
+  const [toggle, setToggle] = useState(false);
+  // const [category, setCategory] = useState(value.activePage);
 
-const Recipes = () => {
-  const classes = useStyles();
-  const maxRecipesShown = 8;
-  let recipes = complexData.results
-    .filter((_, index) => index < maxRecipesShown)
-    .map((recipe, index) => {
-      return (
-        <Grid item xs={12} sm={3} key={index}>
-          <Card className={classes.card}>
-            <CardMedia
-              component="img"
-              alt="food recipe"
-              height="200"
-              className={classes.cardMedia}
-              image={recipe.image}
-              title={recipe.title}
-            />
-            <CardContent className={classes.CardContent}>
-              <Typography gutterBottom title variant="h5">
-                {recipe.title}
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-      );
-    });
+  const handleToggle = () => {
+    console.log("click");
+    setToggle(!toggle);
+  };
 
-  return (
-    <Container className={classes.root}>
-      <Grid container xs={12}>
-        <Grid container xs={6} sm={3}></Grid>
-        <Grid container xs={12} sm={6} direction="row">
-          <Grid item xs={12} sm={9}>
-            <Typography variant="h3" color="textPrimary" gutterBottom>
-              Meals for the week
-            </Typography>
-          </Grid>
-          <Grid item xs={3} sm={6}>
-            <Button
-              variant="contained"
-              color="primary"
-              className={classes.btn}
-              onClick={handleClick}
-            >
-              Fetch Recipes
-            </Button>
-          </Grid>
-        </Grid>
-        <Grid container xs={6} sm={3}></Grid>
-        <Container className={classes.cardGrid}>
-          <Grid container spacing={4} justify="center">
-            {recipes}
-          </Grid>
-        </Container>
-      </Grid>
-    </Container>
-  );
+  const queryItem = state.activeItem.content.replace(/\s/g, "").toLowerCase();
+  const imgUrl = `https://spoonacular.com/cdn/ingredients_100x100/${queryItem}.jpg`;
+  console.log(imgUrl);
+
+  useEffect(() => {
+    const API_ROOT = `https://api.spoonacular.com/`;
+    const findByIngredients = "recipes/findByIngredients?ingredients=";
+    const complexSearch = "recipes/complexSearch";
+    const ingredients = "food/ingredients/search";
+    const numRecipes = 10;
+    const categories = [ingredients, findByIngredients, complexSearch];
+    const URL = `${API_ROOT}${categories[1]}?${queryItem}&number=${numRecipes}&apiKey=${process.env.REACT_APP_SPOONACULAR}`;
+    console.log(URL);
+
+    fetch(URL)
+      .then((res) => {
+        if (res.ok) {
+          return res.json();
+        }
+        throw new Error("Bad Response from Server");
+      })
+      .then((data) => {
+        console.log("got back some data", data);
+        setData(data);
+      })
+      .catch((error) => {
+        console.log("error");
+      });
+
+    return URL;
+  }, [state.activeItem, queryItem]);
+
+  return data === null ? <h1>LOADING</h1> : <RecipeCards data={data} />;
 };
 
 export default Recipes;
-
-// let recipes = recipesArr.map((recipe, index) => {
-//   return (
-//     <div className="recipe-card" key={index}>
-//       <img src={recipe.image} alt={recipe.title}></img>
-//       <h3>{recipe.title}</h3>
-
-//       {/* <p>{recipe.missedIngredients}</p> */}
-//     </div>
-//   );
-// });
