+++ src/styles/index.css
@@ -2,6 +2,8 @@ body {
   margin: 0;
   max-height: 100vmax;
   max-width: 100vmax;
+  height: 100vh;
+  background-color: " #F4A261";
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
     "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
     sans-serif;
